Fixed Windows build.
TBR=dcarney@chromium.org

Review URL: https://codereview.chromium.org/140983005

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@18665 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/extensions/gc-extension.h
@@ -45,7 +45,8 @@ class GCExtension : public v8::Extension {
 
  private:
   static const char* BuildSource(char* buf, size_t size, const char* fun_name) {
-    OS::SNPrintF(Vector<char>(buf, size), "native function %s();", fun_name);
+    OS::SNPrintF(Vector<char>(buf, static_cast<int>(size)),
+                 "native function %s();", fun_name);
     return buf;
   }
